move intrinsic call mangling to the bridge for upstreaming
