+++ flang/include/flang/Optimizer/Support/InternalNames.h
@@ -14,10 +14,6 @@
 #include "llvm/ADT/StringRef.h"
 #include <cstdint>
 
-namespace mlir {
-class FunctionType;
-}
-
 namespace fir {
 
 /// Internal name mangling of identifiers
@@ -126,17 +122,6 @@ private:
   std::string doKinds(llvm::ArrayRef<std::int64_t> kinds);
   std::string toLower(llvm::StringRef name);
 };
-
-/// Returns a name suitable to define mlir functions for Fortran intrinsic
-/// Procedure. These names are guaranteed to not conflict with user defined
-/// procedures. This is needed to implement Fortran generic intrinsics as
-/// several mlir functions specialized for the argument types.
-/// The result is guaranteed to be distinct for different mlir::FunctionType
-/// arguments. The mangling pattern is:
-///    fir.<generic name>.<result type>.<arg type>...
-/// e.g ACOS(COMPLEX(4)) is mangled as fir.acos.z4.z4
-std::string mangleIntrinsicProcedure(llvm::StringRef genericName,
-                                     mlir::FunctionType);
 
 } // namespace fir
